more wxHtmlHelpFrame fixes (didn't store customization correctly, still needs some fixing, will do asap)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4279 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/html/helpfrm.h
@@ -61,7 +61,7 @@ enum
 
 typedef struct
 {
-    long x, y, w, h;
+    int x, y, w, h;
     long sashpos;
     bool navig_on;
 }
